fix: user signup modal
