+++ src/components/Layout.jsx
@@ -2,6 +2,7 @@ import Header from "./Header";
 import Footer from "./Footer";
 import { useLocation } from "react-router-dom";
 import SellerSignUpModal from "./SellerSignUpModal";
+import UserSignUpModal from "./UserSignUpModal";
 
 /**
  * AppLayout component
@@ -21,6 +22,7 @@ const Layout = ({ children }) => {
       <main className="self-center w-full min-h-svh">{children}</main>
       <Footer />
       <SellerSignUpModal />
+      <UserSignUpModal />
     </div>
   );
 };
