--- src/components/SignUpModal/UserSignUpModal.css
@@ -1,29 +1,0 @@
-.custom-checkbox {
-    display: inline-block;
-    width: 24px;
-    height: 24px;
-    background-color: #fff;
-    border: 1px solid #ccc;
-    border-radius: 50%; /* Make it round */
-    position: relative;
-  }
-  
-  .custom-checkbox::after {
-    content: '';
-    display: block;
-    width: 10px;
-    height: 10px;
-    background-color: #fff; /* Initial background color */
-    border-radius: 50%; /* Make the inner circle round */
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%) scale(0); /* Scale down to 0 initially */
-    transition: transform 0.2s ease, background-color 0.2s ease; /* Add transition for animation effect */
-  }
-  
-  .custom-checkbox input[type="checkbox"]:checked + .custom-checkbox::after {
-    background-color: #4a90e2; /* Change background color to primary color when checked */
-    transform: translate(-50%, -50%) scale(1); /* Scale up to 1 when checked */
-  }
-
